feat: Dealer 클래스 추가
참여자와 딜러가 하는 역할이 비슷해서 Player라는 추상 클래스를 만들어서 공통화 처리
