+++ src/main/kotlin/blackjack/domain/Participant.kt
@@ -1,36 +1,17 @@
 package blackjack.domain
 
 data class Participant(
-    val name: String,
-) {
-    val cardList: MutableList<Card> = mutableListOf()
-
-    val score: Int
-        get() {
-            var score = cardList.sumOf { it.rank.point }
-
-            if (cardList.any { it.rank == Rank.ACE } && score <= BLACKJACK_SCORE - ACE_EXTRA_POINT) {
-                score += ACE_EXTRA_POINT
-            }
-            return score
-        }
-
-    val canReceiveCard: Boolean
+    override val name: String
+) : Player(name = name) {
+    override val canReceiveCard: Boolean
         get() {
             return score < 21
         }
 
-    fun receiveCard(card: Card) {
-        cardList.add(card)
-    }
-
     companion object {
-        private const val BLACKJACK_SCORE = 21
-        private const val ACE_EXTRA_POINT = 10
         fun createParticipants(participantString: String): List<Participant> {
             return participantString.split(",")
                 .map { Participant(it.trim()) }
         }
     }
-
 }
